User Story 6, 7  3 Hours
I made it so that events and areas and sale items are all connected and
so that they all show up where they are supposed to. I just didn't get
the artisan on the detail page, but other than that it is good. I'll go
back if I have time.

I downloaded a ton of pictures and fixed the init script as well.