Patch Up Work 2 hours
I worked on making the shopping cart be viewable in the email. It is
hard with Mako to do that, so I made a link to the localhost page
instead of hard coding it.

Then I touched up festivals, so that areas have pictures. And worked on
styling a little bit.